LPS-193924 Inline o component UsersProviderTracker in UserNotificationType, and merge class UsersProviderTracker with UserNotificationType
