+++ modules/apps/notification/notification-service/src/main/java/com/liferay/notification/internal/type/UserNotificationType.java
@@ -9,7 +9,6 @@ import com.liferay.notification.constants.NotificationConstants;
 import com.liferay.notification.constants.NotificationQueueEntryConstants;
 import com.liferay.notification.context.NotificationContext;
 import com.liferay.notification.internal.type.users.provider.UsersProvider;
-import com.liferay.notification.internal.type.users.provider.UsersProviderTracker;
 import com.liferay.notification.model.NotificationQueueEntry;
 import com.liferay.notification.model.NotificationRecipient;
 import com.liferay.notification.model.NotificationRecipientSetting;
@@ -17,6 +16,8 @@ import com.liferay.notification.model.NotificationTemplate;
 import com.liferay.notification.type.BaseNotificationType;
 import com.liferay.notification.type.NotificationType;
 import com.liferay.object.service.ObjectEntryService;
+import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMap;
+import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMapFactory;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.json.JSONUtil;
@@ -31,7 +32,10 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
+import org.osgi.framework.BundleContext;
+import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Deactivate;
 import org.osgi.service.component.annotations.Reference;
 
 /**
@@ -94,9 +98,8 @@ public class UserNotificationType extends BaseNotificationType {
 		NotificationTemplate notificationTemplate =
 			notificationContext.getNotificationTemplate();
 
-		UsersProvider usersProvider =
-			_usersProviderServiceTracker.getUsersProvider(
-				notificationTemplate.getRecipientType());
+		UsersProvider usersProvider = _serviceTrackerMap.getService(
+			notificationTemplate.getRecipientType());
 
 		for (User user : usersProvider.provide(notificationContext)) {
 			if (!_objectEntryService.hasModelResourcePermission(
@@ -147,14 +150,24 @@ public class UserNotificationType extends BaseNotificationType {
 			notificationContext);
 	}
 
+	@Activate
+	protected void activate(BundleContext bundleContext) {
+		_serviceTrackerMap = ServiceTrackerMapFactory.openSingleValueMap(
+			bundleContext, UsersProvider.class, "recipient.type");
+	}
+
+	@Deactivate
+	protected void deactivate() {
+		_serviceTrackerMap.close();
+	}
+
 	@Reference
 	private ObjectEntryService _objectEntryService;
+
+	private ServiceTrackerMap<String, UsersProvider> _serviceTrackerMap;
 
 	@Reference
 	private UserNotificationEventLocalService
 		_userNotificationEventLocalService;
-
-	@Reference
-	private UsersProviderTracker _usersProviderServiceTracker;
 
 }
