--- modules/apps/notification/notification-service/src/main/java/com/liferay/notification/internal/type/users/provider/UsersProviderTracker.java
@@ -1,39 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.notification.internal.type.users.provider;
-
-import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMap;
-import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMapFactory;
-
-import org.osgi.framework.BundleContext;
-import org.osgi.service.component.annotations.Activate;
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Deactivate;
-
-/**
- * @author Feliphe Marinho
- */
-@Component(service = UsersProviderTracker.class)
-public class UsersProviderTracker {
-
-	public UsersProvider getUsersProvider(String recipientType) {
-		return _serviceTrackerMap.getService(recipientType);
-	}
-
-	@Activate
-	protected void activate(BundleContext bundleContext) {
-		_serviceTrackerMap = ServiceTrackerMapFactory.openSingleValueMap(
-			bundleContext, UsersProvider.class, "recipient.type");
-	}
-
-	@Deactivate
-	protected void deactivate() {
-		_serviceTrackerMap.close();
-	}
-
-	private ServiceTrackerMap<String, UsersProvider> _serviceTrackerMap;
-
-}
